# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUrlResponse {
  url: string;
  configured: boolean;
}

export interface ExchangeCodeRequest {
  code: string;
  redirectUri: string;
}

export interface RefreshTokenRequest {
  refreshToken: string;
}

export interface TokenResponse {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
  tokenType: string;
}

export interface AuthStatusResponse {
  configured: boolean;
  geminiConfigured: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface QuoteResponse {
  symbol: string;
  last?: number;
  bid?: number;
  ask?: number;
  change?: number;
  changePct?: number;
  volume?: number;
  high?: number;
  low?: number;
  fiftyTwoWeekHigh?: number;
  fiftyTwoWeekLow?: number;
  peRatio?: number;
  error?: string;
}

export interface Candle {
  datetime: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface PriceHistoryResponse {
  symbol: string;
  candles: Candle[];
  error?: string;
}

export interface OptionContract {
  strike: number;
  expiration: string;
  bid?: number;
  ask?: number;
  last?: number;
  volume?: number;
  openInterest?: number;
  iv?: number;
  delta?: number;
  gamma?: number;
  theta?: number;
  vega?: number;
  dte?: number;
}

export interface OptionChainResponse {
  symbol: string;
  underlyingPrice?: number;
  calls: OptionContract[];
  puts: OptionContract[];
  error?: string;
}

export interface TechnicalAnalysisRequest {
  quote: QuoteResponse;
  candles: Candle[];
  model?: string;
  temperature?: number;
  customPrompt?: string;
}

export interface OptionsAnalysisRequest {
  quote: QuoteResponse;
  chain: OptionChainResponse;
  model?: string;
  temperature?: number;
  customPrompt?: string;
}

export interface ChatRequest {
  question: string;
  marketContext?: string;
  model?: string;
  temperature?: number;
}

export interface AiAnalysisResponse {
  response: string;
  error?: string;
}

export interface ModelsResponse {
  models: string[];
}

export type GetQuoteParams = {
  symbol: string;
  accessToken: string;
};

export type GetPriceHistoryParams = {
  symbol: string;
  accessToken: string;
  timeframe?: string;
};

export type GetOptionChainParams = {
  symbol: string;
  accessToken: string;
  contractType?: string;
  daysToExpiration?: number;
};