# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  question: string;
  marketContext?: string;
  model?: string;
  temperature?: number;
}