# lib/api-zod/src/generated/types/exchangeCodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface ExchangeCodeRequest {
  code: string;
  redirectUri: string;
}