# lib/api-zod/src/generated/types/optionChainResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */
import type { OptionContract } from "./optionContract";

export interface OptionChainResponse {
  symbol: string;
  underlyingPrice?: number;
  calls: OptionContract[];
  puts: OptionContract[];
  error?: string;
}