# lib/api-zod/src/generated/types/optionsAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */
import type { OptionChainResponse } from "./optionChainResponse";
import type { QuoteResponse } from "./quoteResponse";

export interface OptionsAnalysisRequest {
  quote: QuoteResponse;
  chain: OptionChainResponse;
  model?: string;
  temperature?: number;
  customPrompt?: string;
}