# lib/api-zod/src/generated/types/priceHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from "./candle";

export interface PriceHistoryResponse {
  symbol: string;
  candles: Candle[];
  error?: string;
}