# lib/api-zod/src/generated/types/quoteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteResponse {
  symbol: string;
  last?: number;
  bid?: number;
  ask?: number;
  change?: number;
  changePct?: number;
  volume?: number;
  high?: number;
  low?: number;
  fiftyTwoWeekHigh?: number;
  fiftyTwoWeekLow?: number;
  peRatio?: number;
  error?: string;
}