# lib/api-zod/src/generated/types/technicalAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from "./candle";
import type { QuoteResponse } from "./quoteResponse";

export interface TechnicalAnalysisRequest {
  quote: QuoteResponse;
  candles: Candle[];
  model?: string;
  temperature?: number;
  customPrompt?: string;
}