# lib/api-zod/src/generated/types/tokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenResponse {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
  tokenType: string;
}